requirements.txt
groq
streamlit
python-dotenv
yfinance==0.1.70
textblob==0.15.3
ollama